# cpu/schbench: fix schbench test to incorporate the latest updates