This patch is intended to revise the script by integrating the most recent updates from latest schbench (upstream).


**Before Patch:**

```
# avocado run cpu/schbench.py
JOB ID     : e0a0a4c0718efe23a57a0c890c5d52b15a88d85f
JOB LOG    : /root/avocado/job-results/job-2023-11-24T01.42-e0a0a4c/job.log
 (1/1) cpu/schbench.py:Schbench.test: STARTED
 (1/1) cpu/schbench.py:Schbench.test: FAIL: The test failed. Failed command is   /root/avocado/job-results/job-2023-11-24T01.42-e0a0a4c/test-results/1-cpu_schbench.py_Schbench.test/tmp_dirqc2plqkc/1-cpu_schbench.py_Schbench.test/schbench -m 10 -t 10 -p 1000 -r 100 -s 10000  (0.96 s)
RESULTS    : PASS 0 | ERROR 0 | FAIL 1 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 3.54 s

Test summary:
cpu/schbench.py:Schbench.test: FAIL
```

**After Patch:**


```
# avocado run cpu/schbench.py
JOB ID     : 2e3c8a2a34189965dd859fea3c06c0462e133304
JOB LOG    : /root/avocado/job-results/job-2023-11-23T23.55-2e3c8a2/job.log
 (1/1) cpu/schbench.py:Schbench.test: STARTED
 (1/1) cpu/schbench.py:Schbench.test: PASS (6.97 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 18.22 s
```

**Raw Logs:**

```
[stdlog] 2023-11-23 23:55:26,142 avocado.utils.process INFO | Running '  /root/avocado/job-results/job-2023-11-23T23.55-2e3c8a2/test-results/1-cpu_schbench.py_Schbench.test/tmp_dirxjcryda_/1-cpu_schbench.py_Schbench.test/schbench -m 1 -t 1 -p 0 -r 5 -i 5                 -F 256 -n 5 -R 100 -w 0 -L '
 DEBUG| [stderr] Wakeup Latencies percentiles (usec) runtime 5 (s) (64 total samples)
 DEBUG| [stderr] 	  50.0th: 13         (30 samples)
 DEBUG| [stderr] 	  90.0th: 13         (0 samples)
 DEBUG| [stderr] 	* 99.0th: 21         (3 samples)
 DEBUG| [stderr] 	  99.9th: 21         (0 samples)
 DEBUG| [stderr] 	  min=12, max=21
 DEBUG| [stderr] Request Latencies percentiles (usec) runtime 5 (s) (500 total samples)
 DEBUG| [stderr] 	  50.0th: 5288       (0 samples)
 DEBUG| [stderr] 	  90.0th: 5304       (135 samples)
 DEBUG| [stderr] 	* 99.0th: 6888       (38 samples)
 DEBUG| [stderr] 	  99.9th: 8400       (5 samples)
 DEBUG| [stderr] 	  min=5284, max=8396
 DEBUG| [stderr] RPS percentiles (requests) runtime 5 (s) (6 total samples)
 DEBUG| [stderr] 	  20.0th: 99         (6 samples)
 DEBUG| [stderr] 	* 50.0th: 99         (0 samples)
 DEBUG| [stderr] 	  90.0th: 99         (0 samples)
 DEBUG| [stderr] 	  min=99, max=99
 DEBUG| [stderr] current rps: 99.99
 DEBUG| [stderr] Wakeup Latencies percentiles (usec) runtime 5 (s) (65 total samples)
 DEBUG| [stderr] 	  50.0th: 13         (30 samples)
 DEBUG| [stderr] 	  90.0th: 13         (0 samples)
 DEBUG| [stderr] 	* 99.0th: 21         (3 samples)
 DEBUG| [stderr] 	  99.9th: 21         (0 samples)
 DEBUG| [stderr] 	  min=12, max=21
 DEBUG| [stderr] Request Latencies percentiles (usec) runtime 5 (s) (508 total samples)
 DEBUG| [stderr] 	  50.0th: 5288       (0 samples)
 DEBUG| [stderr] 	  90.0th: 5304       (135 samples)
 DEBUG| [stderr] 	* 99.0th: 6888       (38 samples)
 DEBUG| [stderr] 	  99.9th: 8400       (5 samples)
 DEBUG| [stderr] 	  min=5284, max=8396
 DEBUG| [stderr] RPS percentiles (requests) runtime 5 (s) (6 total samples)
 DEBUG| [stderr] 	  20.0th: 99         (6 samples)
 DEBUG| [stderr] 	* 50.0th: 99         (0 samples)
 DEBUG| [stderr] 	  90.0th: 99         (0 samples)
 DEBUG| [stderr] 	  min=99, max=99
 DEBUG| [stderr] average rps: 101.60
```

**Parsed Output**
```
# cat ./test-results/1-cpu_schbench.py_Schbench.test/schbench.json
{
    "wakeup_latencies_percentiles": {
        "percentiles": [
            {
                "percentile_50.0": {
                    "latency": "13",
                    "samples": "30"
                }
            },
            {
                "percentile_90.0": {
                    "latency": "13",
                    "samples": "0"
                }
            },
            {
                "percentile_99.0": {
                    "latency": "21",
                    "samples": "3"
                }
            },
            {
                "percentile_99.9": {
                    "latency": "21",
                    "samples": "0"
                }
            }
        ],
        "min_max": {
            "min": "12",
            "max": "21"
        }
    },
    "request_latencies_percentiles": {
        "percentiles": [
            {
                "percentile_50.0": {
                    "latency": "5288",
                    "samples": "0"
                }
            },
            {
                "percentile_90.0": {
                    "latency": "5304",
                    "samples": "135"
                }
            },
            {
                "percentile_99.0": {
                    "latency": "6888",
                    "samples": "38"
                }
            },
            {
                "percentile_99.9": {
                    "latency": "8400",
                    "samples": "5"
                }
            }
        ],
        "min_max": {
            "min": "5284",
            "max": "8396"
        }
    },
    "rps_percentiles": {
        "percentiles": [
            {
                "percentile_20.0": {
                    "latency": "99",
                    "samples": "6"
                }
            },
            {
                "percentile_50.0": {
                    "latency": "99",
                    "samples": "0"
                }
            },
            {
                "percentile_90.0": {
                    "latency": "99",
                    "samples": "0"
                }
            }
        ],
        "min_max": {
            "min": "99",
            "max": "99"
        }
    },
    "average_rps": 101.6
}
```